Add werewolfTurn into separate folder
