+++ src/utils/turns/index.js
@@ -1,0 +1,1 @@
+export { werewolfTurn } from './werewolfTurn';
